Modif. fonctions

+++ ft_memcpy.c
@@ -2,9 +2,9 @@
 
 void	*ft_memcpy(void *dest, const void *src, size_t howmany)
 {
-	int k;
-	unsigned char *dest_temp;
-	unsigned char *src_temp;
+	size_t			k;
+	unsigned char	*dest_temp;
+	unsigned char	*src_temp;
 
 	dest_temp = (unsigned char *)dest;
 	src_temp = (unsigned char *)src;
